Add files via upload
The file 'RobotVision&Dof.py' contains the code by which your computer sends a detection signal to the Arduino.  In my case, the two classes selected were bottle and cup hence their classIds was used. If you want to change the classes you can do it using coco.names.txt file's classIDs and object list or you can print it. The dog image was to test.

The arduino file(.ino) is for you to flash into the arduino using arduino ide.